backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
ast-grep-py==0.42.0
python-multipart==0.0.20
libclang==18.1.1
pydantic-settings==2.7.0
